Working Settings Menu
Requires testing, graphics dropdown works but not much change and not sure if full screen mode works because the mac is weird so if someone could test it that would be great. You will have to build to test because you cannot do most stuff in the editor. Volume needs sound before it can be done.